Populate user and books.

+++ mylibrary/src/main/java/com/alina/mylibrary/repository/BookUserRepository.java
@@ -1,6 +1,7 @@
 package com.alina.mylibrary.repository;
 
 import com.alina.mylibrary.model.BookUser;
+import com.alina.mylibrary.model.Quizz;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
@@ -11,6 +12,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Repository
 @Transactional
-public interface BookUserRepository extends JpaRepository<BookUser,Long>{
+public interface BookUserRepository extends JpaRepository<BookUser,Integer>{
 
 }
+
